requirements.txt
spaces
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
pytorch-lightning==2.1.2
einops
omegaconf
deepspeed
torchmetrics
webdataset
accelerate
tensorboard
PyMCubes
trimesh
rembg
transformers
diffusers==0.28.2
bitsandbytes
imageio[ffmpeg]
xatlas
plyfile
xformers==0.0.22.post7
git+https://github.com/NVlabs/nvdiffrast/
huggingface-hub
gradio_client >= 0.12
rerun-sdk>=0.16.0,<0.17.0
gradio_rerun
